pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["FabianusArdian <fabianusardianu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-jwt-extended = "^4.6.0"
poetry-dotenv-plugin = "^0.2.0"
flask-sqlalchemy = "^3.1.1"
sqlalchemy-orm = "^1.2.10"
werkzeug = "^3.0.6"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.1.0"
flasgger = "^0.9.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
